Migrate reset password endpoint, think of an approach for rate-limited endpoints
