+++ server/service/transport_users.go
@@ -13,11 +13,3 @@ func decodeForgotPasswordRequest(ctx context.Context, r *http.Request) (interfac
 	}
 	return req, nil
 }
-
-func decodeResetPasswordRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var req resetPasswordRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		return nil, err
-	}
-	return req, nil
-}
